requirements.txt
transformers==5.0.0
torch==2.10.0
fastapi==0.128.0
uvicorn==0.40.0
accelerate==1.12.0
requests>=2.31.0
pydantic>=2.7.0
numpy>=1.17
diffusers>=0.25.0
pillow>=10.0.0
python-dotenv>=1.0.0
